rebalancing_adjustment_for_exact_out
