+++ contracts/transmuter/src/swap/non_alloyed_exact_amount_in.rs
@@ -1,13 +1,10 @@
-use std::cmp::Ordering;
-
 use cosmwasm_std::{
-    coin, ensure, to_json_binary, Addr, BankMsg, Coin, Deps, DepsMut, Int256, Response, Uint128,
+    ensure, to_json_binary, Addr, BankMsg, Coin, Deps, DepsMut, Int256, Response, Uint128,
 };
 
 use crate::{
-    asset::convert_amount,
     contract::Transmuter,
-    swap::common::{Adjustment, SwapExactAmountInResponseData},
+    swap::{adjust_exact_in, common::SwapExactAmountInResponseData},
     transmuter_pool::TransmuterPool,
     ContractError,
 };
@@ -21,6 +18,10 @@ impl Transmuter {
         mut deps: DepsMut,
     ) -> Result<Response, ContractError> {
         let pool = self.pool.load(deps.storage)?;
+        let std_norm_factor = pool.std_norm_factor()?;
+        let token_out_norm_factor = pool
+            .get_pool_asset_by_denom(token_out_denom)?
+            .normalization_factor();
 
         let run_pool = |deps: Deps, pool: TransmuterPool| {
             self.out_amt_given_in(deps, pool, token_in, token_out_denom)
@@ -28,40 +29,12 @@ impl Transmuter {
 
         let rebalancing_adjustment =
             |pool: TransmuterPool, token_out: Coin, total_adjustment_value: Int256| {
-                let std_norm_factor = pool.std_norm_factor()?;
-                let token_out_norm_factor = pool
-                    .get_pool_asset_by_denom(token_out_denom)?
-                    .normalization_factor();
-
-                let (token_out, adjustment) = match total_adjustment_value.cmp(&Int256::zero()) {
-                    Ordering::Less => {
-                        // deduct fee from token_out
-                        let fee_amount = convert_amount(
-                            total_adjustment_value.abs().try_into()?,
-                            std_norm_factor,
-                            token_out_norm_factor,
-                            // rounding up means slightly more fee than required, keeps the incentive pool healthy
-                            &crate::asset::Rounding::Up,
-                        )?;
-                        let token_out_amount: Uint128 = token_out.amount.checked_sub(fee_amount)?;
-                        let token_out = coin(token_out_amount.u128(), token_out.denom.clone());
-                        let token_out_denom = token_out.denom.clone();
-
-                        (
-                            token_out,
-                            Adjustment::DeductFee {
-                                fee: coin(fee_amount.u128(), token_out_denom),
-                            },
-                        )
-                    }
-                    Ordering::Greater => (
-                        token_out,
-                        Adjustment::CreditIncentive {
-                            incentive: total_adjustment_value.abs().try_into()?,
-                        },
-                    ),
-                    Ordering::Equal => (token_out, Adjustment::None),
-                };
+                let (token_out, adjustment) = adjust_exact_in(
+                    token_out,
+                    token_out_norm_factor,
+                    std_norm_factor,
+                    total_adjustment_value,
+                )?;
 
                 ensure!(
                     token_out.amount >= token_out_min_amount,
