+++ contracts/transmuter/src/swap/tokens_to_alloyed_asset.rs
@@ -1,15 +1,16 @@
-use std::cmp::Ordering;
-
 use cosmwasm_std::{coin, ensure, Addr, Coin, Deps, DepsMut, Env, Int256, Response, Uint128};
 use osmosis_std::types::osmosis::tokenfactory::v1beta1::MsgMint;
 
 use crate::{
     alloyed_asset::swap_to_alloyed,
-    asset::convert_amount,
     contract::Transmuter,
-    swap::common::{
-        set_data_if_sudo, Adjustment, Entrypoint, SwapExactAmountInResponseData,
-        SwapExactAmountOutResponseData,
+    swap::{
+        adjust_exact_out,
+        common::{
+            adjust_exact_in, set_data_if_sudo, Adjustment, Entrypoint,
+            SwapExactAmountInResponseData, SwapExactAmountOutResponseData,
+        },
+        rebalancing_adjustment_for_exact_out,
     },
     transmuter_pool::TransmuterPool,
     ContractError,
@@ -130,36 +131,12 @@ impl Transmuter {
                 let std_norm_factor = pool.std_norm_factor()?;
                 let token_out_norm_factor = alloyed_norm_factor;
 
-                let (token_out, adjustment) = match total_adjustment_value.cmp(&Int256::zero()) {
-                    Ordering::Less => {
-                        // deduct fee from token_out
-                        let fee_amount = convert_amount(
-                            total_adjustment_value.abs().try_into()?,
-                            std_norm_factor,
-                            token_out_norm_factor,
-                            // rounding up means slightly more fee than required, keeps the incentive pool healthy
-                            &crate::asset::Rounding::Up,
-                        )?;
-
-                        let token_out_amount: Uint128 = token_out.amount.checked_sub(fee_amount)?;
-                        let token_out = coin(token_out_amount.u128(), token_out.denom.clone());
-                        let token_out_denom = token_out.denom.clone();
-
-                        (
-                            token_out,
-                            Adjustment::DeductFee {
-                                fee: coin(fee_amount.u128(), token_out_denom),
-                            },
-                        )
-                    }
-                    Ordering::Greater => (
-                        token_out,
-                        Adjustment::CreditIncentive {
-                            incentive: total_adjustment_value.abs().try_into()?,
-                        },
-                    ),
-                    Ordering::Equal => (token_out, Adjustment::None),
-                };
+                let (token_out, adjustment) = adjust_exact_in(
+                    token_out,
+                    token_out_norm_factor,
+                    std_norm_factor,
+                    total_adjustment_value,
+                )?;
 
                 ensure!(
                     token_out.amount >= token_out_min_amount,
@@ -215,6 +192,7 @@ impl Transmuter {
         let pool: TransmuterPool = self.pool.load(deps.storage)?;
         let response = Response::new();
 
+        let std_norm_factor = pool.std_norm_factor()?;
         let token_in_norm_factor = pool
             .get_pool_asset_by_denom(token_in_denom)?
             .normalization_factor();
@@ -231,62 +209,12 @@ impl Transmuter {
             Ok((pool, token_in))
         };
 
-        let rebalancing_adjustment =
-            move |pool: TransmuterPool, token_in: Coin, total_adjustment_value: Int256| {
-                let std_norm_factor = pool.std_norm_factor()?;
-                let token_in_norm_factor = pool
-                    .get_pool_asset_by_denom(&token_in_denom)?
-                    .normalization_factor();
+        let rebalancing_adjustment = rebalancing_adjustment_for_exact_out(
+            token_in_max_amount,
+            std_norm_factor,
+            token_in_norm_factor,
+        );
 
-                // If adjustment value is negative, fee take from the token_in, so we require addtional token_in to pay for the fee.
-                // Otherwise, return the token_in as is
-                let (token_in, adjustment) = match total_adjustment_value.cmp(&Int256::zero()) {
-                    // negative adjustment value means fee deduction from token_in
-                    Ordering::Less => {
-                        let fee = convert_amount(
-                            total_adjustment_value.abs().try_into()?,
-                            std_norm_factor,
-                            token_in_norm_factor,
-                            // rounding up means slightly more fee than required, keeps the incentive pool healthy
-                            &crate::asset::Rounding::Up,
-                        )?;
-
-                        let token_in_amount = token_in.amount.checked_add(fee)?;
-
-                        (
-                            coin(token_in_amount.u128(), token_in.denom.clone()),
-                            Adjustment::DeductFee {
-                                fee: coin(fee.u128(), token_in.denom.clone()),
-                            },
-                        )
-                    }
-                    // positive adjustment value means incentive credit to the beneficiary
-                    Ordering::Greater => (
-                        token_in,
-                        Adjustment::CreditIncentive {
-                            incentive: total_adjustment_value.abs().try_into()?,
-                        },
-                    ),
-                    // zero adjustment means no adjustment
-                    Ordering::Equal => (token_in, Adjustment::None),
-                };
-
-                let token_in_amount = token_in.amount.clone();
-
-                ensure!(
-                    token_in_amount <= token_in_max_amount,
-                    ContractError::ExcessiveRequiredTokenIn {
-                        limit: token_in_max_amount,
-                        required: token_in_amount,
-                    }
-                );
-
-                Ok((token_in, adjustment))
-            };
-
-        // if it's exact_out
-        // - fee case: we increase tokens in requirement, it's spreaded through all the tokens in
-        // - incentive case: we credit incentive to the beneficiary
         let (pool, token_in, _adjustment) = self.rebalancer_pass(
             deps.branch(),
             pool,
